refactor environment handling to independent class/files
